Tratamento de Excecao

+++ AppContatos/src/main/java/br/com/ehmf/AppContatos/AppContatos/service/PessoaService.java
@@ -6,6 +6,8 @@ import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import br.com.ehmf.AppContatos.AppContatos.exception.BadRequestException;
+import br.com.ehmf.AppContatos.AppContatos.exception.ResourceNotFoundException;
 import br.com.ehmf.AppContatos.AppContatos.model.Pessoa;
 import br.com.ehmf.AppContatos.AppContatos.repository.PessoaRepository;
 import br.com.ehmf.AppContatos.AppContatos.service.interfaces.PessoaServiceInterface;
@@ -23,15 +25,23 @@ public class PessoaService implements PessoaServiceInterface {
 
 	@Override
 	public Pessoa criar(Pessoa pessoa) {
+		if (pessoa.getId() != null)
+		{
+			throw new BadRequestException("[Criar Pessoa] - id deve ser nulo");
+		}
 		return pessoaRepository.save(pessoa);
 	}
 
 	@Override
 	public Pessoa atualizar(Pessoa pessoa) {
+		if (pessoa.getId() == null)
+		{
+			throw new BadRequestException("[Atualizar Pessoa] - id não pode ser nulo");
+		}
 		Optional<Pessoa> consultaPessoa = pessoaRepository.findById(pessoa.getId());
 		if (!consultaPessoa.isPresent())
 		{
-			return null;
+			throw new ResourceNotFoundException("[Atualizar Pessoa] - pessoa " + pessoa.getId() + " não encontrada");
 		}		
 		Pessoa atualizaPessoa = consultaPessoa.get();
 		atualizaPessoa.setNome(pessoa.getNome());
@@ -44,11 +54,21 @@ public class PessoaService implements PessoaServiceInterface {
 
 	@Override
 	public void excluir(Long idPessoa) {
+		Optional<Pessoa> excluirPessoa = pessoaRepository.findById(idPessoa);
+		if (!excluirPessoa.isPresent())
+		{
+			throw new ResourceNotFoundException("[Excluir Pessoa] - pessoa " + idPessoa+ " não encontrada");
+		}	
 		pessoaRepository.deleteById(idPessoa);		
 	}
 
 	@Override
 	public Optional<Pessoa> consultar(Long idPessoa) {
+		Optional<Pessoa> consultarPessoa = pessoaRepository.findById(idPessoa);
+		if (!consultarPessoa.isPresent())
+		{
+			throw new ResourceNotFoundException("[Excluir Pessoa] - pessoa " + idPessoa+ " não encontrada");
+		}	
 		return pessoaRepository.findById(idPessoa);
 	}
